React to OBS state changes with MIDI messages
This will add a "bidirectional" option for buttons that change the
current or preview scene. When we see an update from OBS that the scene
changed we try to inform the MIDI controls that are associated with that
scene about it if it's bidirectional. In other words: If your scene
changing button has an LED it will turn on/off.
I tested this with a KORG nanoKONTROL2 where it works beautifully.