Separate in and out ports
If the same midi device can be used for input and output, or if it's an
ioport, then the "bidirectional" key in the config should be enough, if
you have a separate device for output, then add it as a 2nd device and
add "out_deviceID": <deviceID> to the config. It's not done in setup.py
for you.